typecd: Make peripheral PD ID update command protected
We want derived classes to be able to trigger an update of the PD
Identity information when they want (instead of only when the object is
created). To enable this, make the UpdatePDIdentityVDOs() function
protected.

Also update the description text to resolve some of the issues which are
no longer outstanding.

BUG=b:174057058
TEST=Run the typecd unit tests. Ensure they all still pass.

Change-Id: I1e5093e76924a8d074849215a4bf8b78c748291e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2555685
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Reviewed-by: Prashant Malani <pmalani@chromium.org>
Commit-Queue: Prashant Malani <pmalani@chromium.org>
Tested-by: Prashant Malani <pmalani@chromium.org>