sirenia: Refactor sirenia library
Refactor the sirenia package to put most of the code in libsirenia with
sirenia itself only containing trichechus and dugong specific code. This
allows for code reuse especially in the interfaces with the rest of the
system, namely manatee-client and manatee-runtime.

BUG=b:173598376
TEST=cargo build
cargo test

Cq-Depend: chromium:2561034
Change-Id: Id70cb74ea2bbf0383790b5eaf3c60fff616b4194
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2560828
Tested-by: Nicole Anderson-Au <nvaa@google.com>
Commit-Queue: Allen Webb <allenwebb@google.com>
Reviewed-by: Allen Webb <allenwebb@google.com>