typecd: Add getter for peripheral syspath
There are circumstances where the derived class (e.g a Cable) may need
to access the syspath of the periheral. For example, on object init, a
Partner may want to look through the syspath for any existing alt modes.

To enable this, make the syspath accessible to derived classes via a
getter.

BUG=b:174057058
TEST=Run the typecd unit tests. Ensure they all still pass.

Change-Id: Iff8844d5411cfb5ac450d799e56805257b150d3e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2559152
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Reviewed-by: Prashant Malani <pmalani@chromium.org>
Commit-Queue: Prashant Malani <pmalani@chromium.org>
Tested-by: Prashant Malani <pmalani@chromium.org>