biod: Use std::unique_ptr for template to avoid copying
Previously a base::Optional is used for retrieving template from
CrosFpDevice. That causes biod to crash at the end of enrollment
because the entire template is too big, base::Optional would
make a copy of the brillo::SecureVector, and brillo::SecureAllocator
cannot allocate such a bug chunk of memory at once. Specifically it
cannot mlock such a bug buffer:
https://chromium.googlesource.com/chromiumos/platform2/+/a68b2fcc897ab6f19ca5434f4ab859a8e7ee2196/libbrillo/brillo/secure_allocator.h#115

Use a smart pointer so that no copying of the template is needed.

BUG=b:173655013, b:172906334
TEST=Unittests
TEST=Enrollment succeeds on Nocturne

Change-Id: I53d569b1159913f2ba9636d90004ad0b208cd6b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2565146
Tested-by: Yicheng Li <yichengli@chromium.org>
Commit-Queue: Yicheng Li <yichengli@chromium.org>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>