vm_tools: Use a separate dbus error buffer for SetHints
Per https://dbus.freedesktop.org/doc/api/html/group__DBusErrors.html "An
error passed to a D-Bus function must always be unset; you can't pass in
an error that's already set".
We ignore errors in SetHints and use a shared field for dbus errors
which means that an error in SetHints results in an already-set error
being passed in, tripping an assertion in libdbus and crashing Garcon.
We fix this by using a separate ScopedDBusError for SetHints which takes
care of initialisation and freeing after it goes out of scope.

Other cases in package_kit_proxy look fine, since in other cases dbus
errors fail the operation, deleting this and wiping out the dbus error
buffer as part of the destructor.

BUG=chromium:1050830
TEST=CQ

Change-Id: Iabbcbc7e2e80e31a323d1ee8f2f26da760f09c38
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2567270
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Reviewed-by: David Munro <davidmunro@google.com>
Tested-by: David Munro <davidmunro@google.com>
Commit-Queue: David Munro <davidmunro@google.com>