shill-dbus-client: Add service type to change handler.
This patch adds the type property of the default service to the
OnDefaultServiceChanged handler. This is needed in order to be able to
discover when the service is a VPN for example as the Device type does
not always provide this unambiguously.

Note this approach can be revisited/reverted if/when the device does
provide this level of detail and/or a service abstraction is defined and
passed back.

BUG=b:173065426
TEST=units

Change-Id: Ifc175d610f2d63ebba0ee78b164f09540d3f792d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2569159
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Garrick Evans <garrick@chromium.org>
Commit-Queue: Garrick Evans <garrick@chromium.org>