cryptohome: simplify the timestamp file interaction
Instead of dealing with markers of ts file existence in the keyset, just
unconditionally read it on the load path and just use the legacy field
if the file is missing. This CL is a preparation to the untying
timestamp from the keyset entirely by introducing a single ts file (in
contrast with current per-keyset ones).

BUG=b:172344610
TEST=unittest

Change-Id: I6f0273005e83198e8561664c36dfa0b9c1443e19
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2569474
Commit-Queue: Daniil Lunev <dlunev@chromium.org>
Tested-by: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>