arc: arcvm-mount-myfiles: Bind-mount MyFiles/Downloads/ only if needed
The Downloads directory is going to be a standard subdirectory of the
MyFiles directory instead of a bind-mount from its sibling.
This CL modifies the way to prepare /home/chronos/user/MyFiles/Downloads
in the concierge namespace so that it will be a bind-mount only when it
is actually a bind-mount in the init namespace.

BUG=b:174727521
TEST=/run/arcvm/media/MyFiles is properly set up regardless of whether
TEST=(1) /home/chronos/user is a mount point in the concierge namespace,
TEST=(2) /home/chronos/user/MyFiles/Downloads is a mount point in the
TESST=   init namespace.

Change-Id: Ie2045a08af483312be747e0bd8ef0de5a875be76
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2570884
Tested-by: Youkichi Hosoi <youkichihosoi@chromium.org>
Reviewed-by: Satoshi Niwa <niwa@chromium.org>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Commit-Queue: Youkichi Hosoi <youkichihosoi@chromium.org>