shill: avoid backward state transition during EAP re-key
On 802.1X networks, the AP may re-initiate an EAP authentication. This
causes unnecessary state transitions, and as seen in $BUG, we can get
stuck in the Associating state.

Add a flag that is switched on when we see an EAP authentication start
so that we can avoid transitioning away from the Online state.

BUG=b:174158532
TEST=unit tests
TEST=wifi_matfunc

Change-Id: I7dfb0a2068315c3338956ef62ea70fbe81c4c703
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2573458
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Matthew Wang <matthewmwang@chromium.org>