cryptohome: Start consolidating file paths definitions.
Currently the paths are put all over the place. Particularly shadow root
is passed to most storage related components, but is pretty much never
overridden (and it doesn't need to be because of the fake platform now).
It makes sense to consolidate it in one place to be accessed statically
instead of keeping track of it in tons of places. This CL also moves a
bunch of other path definitions to keep it compilable.

BUG=chromium:1124561
TEST=unittest
TEST=tast run <DUT> hwsec.*

Change-Id: Ib350752ecf5d9623be3dfe15bc09b617f17f4fbb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2574144
Tested-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>