biod: FpFrameCommand should copy payload after running command
FpFrameCommand was copying *Resp() into the received FpFramePacket
before the command is even run. This causes FpFrameCommand to
receive all zeros in the template (the size is correct). This
in turn causes the template data written to storage to be AAAAA...
(base64 encode of zeros). This in turn causes the template version
(received by firmware) to be zero. This inturn causes biod to detect
a template version mismatch when loading records and delete the record.

BUG=b:174790814, b:174852354
TEST=On Nocturne, enroll -> match -> signout/signin -> match success
TEST=Template record data is no longer full of AAAAs in daemon store
TEST=Modified unittest catches this bug if fp_frame_command.cc is
     reverted.

Change-Id: Id52f76e625f19e7a37d890cd78571e42306e014b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2575716
Tested-by: Yicheng Li <yichengli@chromium.org>
Commit-Queue: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>