shill: flip is_rekey_in_progress_ switch on backward state transition
CL:2573458 captured the re-key case for 802.1X networks, but not for PSK
networks. Assume all supplicant state transitions from completed to an
auth/assoc state indicate that a re-key is in progress (instead of only
when receiving an EAP start event).

BUG=b:174158532
TEST=updated wifi.PTK and wifi.GTK to ensure that shill never
transitions out of a Connected state

Change-Id: Ib4e4d26b439c8b960d5e90e1ece6b83df7ceb2f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2576041
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>