system_api: Add ListKeysEx constant for cryptohome
Add a constant for the ListKeysEx D-Bus method exposed by the cryptohome
daemon. This allows to use this method in the Chromium repository
without hardcoding the name of the method in the code.

This method has been exposed since ever, but there's been no constant
for it in the system_api library.

BUG=chromium:1104120
TEST=build completes

Change-Id: I6dd08a3c101a7f6c308af0c76dcc7e5a0543a213
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2574982
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Leo Lai <cylai@google.com>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Leo Lai <cylai@google.com>
Reviewed-by: Leo Lai <cylai@google.com>