fingerprint_study: Add check for writing empty files
Since the files this function can produce might be gibberish to the
person using this tool, we need to ensure that we give the user
enough signal that something is wrong. Here we simply don't produce
the file if the size is obviously incorrect.
Writing zero size files occurred when gnupg was incorrectly configured.

BUG=b:144061947
TEST=Checked on host with bad pub key file

Change-Id: I2148da0e288675d9cb4d4fff516c828ced8fb4f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2579685
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Commit-Queue: Craig Hesling <hesling@chromium.org>
Tested-by: Craig Hesling <hesling@chromium.org>