diagnostics: add cancel functionality to memory routine
The memory routine runs the memtester executable, which can take
a very long time and leave the device sluggish while it's running.
Add the ability to cancel the memory routine (which is just killing
the memtester executable). This way, if a user decides they no
longer want to wait for the routine to finish, they don't have to
wait half an hour or so to use their device.

This requires some changes to cros_healthd's executor to allow it
to track processes it has started. Currently, processes are tracked
by the path to their executable, so only one process can be tracked
per executable. For example, two instances of the memory routine
cannot be ran concurrently. This is an acceptable tradeoff for now,
since getting the cancel functionality in is high priority, and it
doesn't make sense to have two memtester executables running
concurrently anyway - the first will mlock all the device's memory
and the second will have nothing to do. This could be revisited if
we want to track processes in the future which could run the same
executable concurrently.

BUG=b:173062671
TEST=cros-health-tool diag --action=run_routine --routine=memory
--force_cancel_at_percent=3 &
ps -aux | grep memtester (shows cros_healthd running memtester)
wait for routine to be cancelled at 3%, then:
ps -aux | grep memtester (no memtester processes running)

Change-Id: Ibd5119cc1b55ee07cdbe0b355e698bae512bf24f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2579984
Tested-by: Paul Moy <pmoy@chromium.org>
Commit-Queue: Paul Moy <pmoy@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>