lorgnette: Make PortToken uncopyable
When base::Optional is created from a PortToken during
GetCapabilities(), it calls the copy constructor by default.  The
temporary copy then immediately releases the port when it goes out of
scope.  Remove the copy constructor and add a move constructor instead
so that port ownership can only be passed on instead of duplicated.
This should only affect pixma network scanners that don't support eSCL
because that is the only case where we were requesting port access.

BUG=b:175143088
TEST=lorgnette_cli with a pixma scanner

Change-Id: If4a37d3add7d19d674c87058d91286c2a5c4da98
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2582514
Reviewed-by: Jesse Schettler <jschettler@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>