REVERT-ME: power: Send upstart events for modem WDS status
This is a temporary hack for Modem SSR on trogdor and will be
removed once qc-netmgr is merged into the driver.
As per QC WDS is a better indicator of the health of modem.
During a modem crash, the node can still be up, but WDS service
will be offline. Here we make sure WDS and the Node are up before
sending a qrtr-service-added message. Similarly
qrtr-service-removed messages are sent when WDS service goes offline.
These messages are used to start/stop qc-netmgr.

BUG=b:159860147, b:172173518
TEST=cros deploy ${DUT} power_manager && reboot && diag-route && send_data 75
37 03 00 00

Change-Id: I7cd99cb5f8b6738bca10e9a1962bcb200d10be4d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2582515
Tested-by: Madhav . <madhavadas@google.com>
Commit-Queue: Madhav . <madhavadas@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Madhav . <madhavadas@google.com>