shill: add WiFi.RoamState D-Bus property
Due to CL:2427764, we no longer transition away from the Online state
during a roam. However, many tests wait on certain state changes to
verify that the DUT successfully connects to a Service. Expose a roam
state via D-Bus that tracks the actual state of the Service so that we
can still maintain the Online (or other Connected) state.

BUG=b:171086223
TEST=unit tests
TEST=deploy to device, run wifi.RoamDbus, and monitor for the RoamState
property

Change-Id: Ib7c927717a83af8296d7461a7fefb8dcc2aabeaa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2582450
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>