shill: Fix registering of callbacks in DBusPropertiesProxy
A bug in the DBusPropertiesProxy implementation was exposed when
introducing asynchronous calls to short lived proxies.

ChangedSignalHandlers should only be registered when a callback
is provided, otherwise a call to a null callback will be attempted
(and the change event would have no purpose).

This fixes that by only registering change signal handlers when a
callback is set.

BUG=b:175042773
TEST=No crashes :-)

Change-Id: I4fc120365d9497cefb7c3b412cc06b8129bcdcdf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2587868
Tested-by: Steven Bennetts <stevenjb@chromium.org>
Auto-Submit: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Steven Bennetts <stevenjb@chromium.org>