shill: drop Mac80211Monitor
This monitor's only real work is to poke at
/sys/kernel/debug/ieee80211/<phyX>/wake_queues if it detects a queue
hang for more than 60 seconds. This seems like a bit of a driver hack,
and it's not present on any supported kernel now; it was a non-upstream
feature added only for some old Atheros chipsets which are no longer
supported.

This module also logs and reports some metrics, even if the
'wake_queues' mechanism isn't available, but we don't really use those
except to monitor the 'wake_queues' feature.

So, this code is no longer useful. Drop it.

BUG=b:172219309
TEST=unit tests

Change-Id: I4b62f663e506d5a7c25be8c5e68b6d9d7fe5ff0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2590497
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>