patchpanel: Resolve deadlock in reading stdout pipe
Currently, MinijailedProcessRunner reads the stdout pipe after the
execution of subprocess is done. This may cause a deadlock when the
buffer size of the pipe cannot accommodate the whole output: subprocess
is stuck at writing to the pipe since the buffer is full, while
patchpanel is waiting for the execution finish.

This patch resolves this deadlock by moving the logic for reading the
pipe before waitpid().

Note that after this patch, patchpanel may still be not responsive if
the subprocess itself running by this runner is stuck, or takes too much
time.

BUG=b:175842822
TEST=deployed on octopus, add 4000 rules in the mangle table by `for i
  in {1..4000}; do iptables -t mangle -A tx_eth0 -j ACCEPT; done`,
  invoke the traffic counters by `in {1..1000}; do dbus-send --system
  --dest=org.chromium.PatchPanel --print-reply /org/chromium/PatchPanel
  org.chromium.PatchPanel.GetTrafficCounters; done` (patchpanel code is
  modified to accept requests with no parameter), after that, checked
  that no fd is leaked and no suspicious log in /var/log/net.log

Change-Id: I740fa40b83c39ff4510abe7f4e93afe00ee43217
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2593504
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>