patchpanel: Avoid invoking Manager::OnShutdown() twice
Manager::OnShutdown() will be called "when the message loop exits and
before Daemon::Run() returns", as mentioned in the comment of
brillo::Daemon::OnShutdown(), so we don't need to call it in ~Manager().

Also add the call to the base implementation in Manager::OnShutdown(),
which is asked by Daemon::OnShutdown().

BUG=None
TEST=Ran `stop patchpanel`, saw "Shutting down and cleaning up" exactly
  once in the log.

Change-Id: If486104060131c36117f8226d128fd1737d6aff1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2595751
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>