libchromeos-rs: Add chromeos module that depends on system_api-rust.
This adds Chrome OS specific code behind an optional feature so that
the extra dependencies can be gated behind the feature.

The module adds functionality for getting the user id hash, the
daemon-store path for a given daemon, and whether the device is in
developer mode or not.

BUG=chromium:1163664
TEST=cargo test --all-features

Cq-Depend: chromium:2611405
Change-Id: Ic1e76cf2a42148bf8d6eab3f3489ed62bcd66d74
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2598142
Tested-by: Allen Webb <allenwebb@google.com>
Commit-Queue: Allen Webb <allenwebb@google.com>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>