crosh: network_diag: drop --wifi-mon, ath9k
This option only worked for ath9k chipsets, which we don't support any
more. We could probably support some of this (e.g., the 'station dump')
for other chips, but this is a lot of shell code, and apparently nobody
wanted to be using it for non-ath9k so far.

BUG=none
TEST=other network_diag options still work

Change-Id: Iaf1edafae165d8230c20758f4ed93fb2ed647545
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2599449
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>