crash: Move kernel_collector to RE2.
This is mostly a trivial change, EXCEPT for the BIOS's banner_re. That
regex originally used positive lookahead (to match the crash message
without capturing it), but RE2 does not support that feature. The
specific use of positive lookahead was intended to allow the regex to
skip past the text before the crash message, so instead that is
reimplemented manually.

BUG=chromium:1133453
TEST=unit tests

Change-Id: I0a925f9385c6d04c8009b9c87f4cbb5de0235836
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2600443
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Michael Irani <michaelirani@google.com>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>