u2fd: Add dbus method to test legacy u2f credentials
Add HasLegacyCredentials dbus method. This is to be used by the
browser for supporting signing legacy credentials registered via
the U2F interface. If the browser notices that no fingerprint or
PIN is setup for WebAuthn, but power button mode is allowed, the
browser will call this method with a list of credentials. If u2fd
says some credentials match, the browser will dispatch GetAssertion
requests to u2fd, with verification_type=verification_discouraged,
to indicate this should be signed with physical presence.

BUG=b:165724836
TEST=unittests

Change-Id: I00edb3385f294ed2a4621b313a1e7f3f9aa95c8a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2600446
Tested-by: Yicheng Li <yichengli@chromium.org>
Commit-Queue: Yicheng Li <yichengli@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>