typecd: Implement Cable DiscoveryComplete command
In order to run mode entry code, the PortManager needs to verify that
cable discovery parsing is complete. Implement the DiscoveryComplete
function for the Cable class and have the PortManager call it (via the
Port object) when it runs the mode entry code.

Also, update the CableTest code to:
- Use ScopedTempDir, which ensures the temp files are cleaned up after
  the test execution.
- Correct the SOP' plug alt mode paths to have a prefix for the SOP'
  plug device; this will lay the groundwork to update the test to check
  for "number_of_alternate_modes" once the patches which implement that
  parsing are submitted.

BUG=b:172097194
TEST=Unit tests still pass.

Change-Id: I98e4c33ff91b341cc2c3f38b7678d22ab4e315dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2601504
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Reviewed-by: Prashant Malani <pmalani@chromium.org>
Commit-Queue: Prashant Malani <pmalani@chromium.org>
Tested-by: Prashant Malani <pmalani@chromium.org>