camera: hal_adapter: Add JPEG orientation to ZSL reprocessing request
Adds JPEG orientation to the ZSL private reprocessing request. This is
needed because the original preview capture requests don't come with the
tag and we need to manually add it back when doing private reprocessing
so that the resulting JPEG is of the correct orientation.

BUG=b:176414732
TEST=Take picture on dedede with all 4 orientations.

Change-Id: Iecd61dc2274d538eb9d979ae23ab7ffd4851be82
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2605040
Reviewed-by: Hung-yu Wu <hywu@chromium.org>
Reviewed-by: Ren-Pei Zeng <kamesan@chromium.org>
Tested-by: Jasmine Chen <lnishan@google.com>
Commit-Queue: Jasmine Chen <lnishan@google.com>