mount-encrypted: `Persist()` work with the new `kWritable` state.
For TPM1.2, the encstateful space is not considered to be writable if
the content is not initialized. However when the space is created with
correct PCR binding and attributes, the write should succeed.

For testing, the following OOBE cases for TPM1.2 are tested while
TPM2.0 is built successfully:
1. when encstateful space is not defined, trivially system key is
selected as usual.
2. when encstateful space is defined but no value is written, the
encstateful space is selected, which is the case this CL is trying to
address.
3. when encstateful space contains the system key from last OOBE, the
encstateful space is selected and the system key is re-generated as
usual.
4. when preservation request, the encstateful space is selected as
usual. Note that this is emulated by hard-coding the firmware update
request in mount-encrypted.

BUG=b:168672088
BUG=b:176463814 # meta bug of TPM init flow.
TEST=See above.

Change-Id: I5790f44cdbac3b07e74064b8d656d57a7c5121ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2607032
Reviewed-by: Mattias Nissler <mnissler@chromium.org>
Tested-by: Leo Lai <cylai@google.com>
Commit-Queue: Leo Lai <cylai@google.com>