pciguard: Add Intel Giga Bit Ethernet driver to allowlist
Need: To ease validation, development, certification.
Instead of manually adding the driver either through echo or init script,
having included in the allowlist will ease validation and
certification a lot.

Please note this is temporarily added to the allowlist.
This needs to be revisited and reviewed for final inclusion.
Refer to: b:163121310.

BUG=b:151065827,b:163121310,b:173775390
TEST=Verif that igb is included the in the allowlist.
     cat /sys/bus/pci/drivers_allowlist
     Verify that TBT devices with igb are enumerated automatically.

Signed-off-by: Chiranjeevi Rapolu <chiranjeevi.rapolu@intel.corp-partner.google.com>
Change-Id: I026d651924a0ab45728e9e5bf099c4e7a5abc2c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2613308
Commit-Queue: Rajat Jain <rajatja@google.com>
Reviewed-by: Aashay Shringarpure <aashay@google.com>