installer: Migrate flag processing of cros_installer to brillo:FlagHelper
This makes it easier to add flags related to chromeos-install here
easily.

Since we call chromeos-postinst of the newly installed image as a
mounted file system, it is fine to edit this file without breaking
anything (as long as it is still a shell script).

BUG=b:176492189
TEST=chromeos-install # through USB
TEST=cros flash # of newly built image.

Change-Id: Ia84da90a41f01ad4e4cdd57e699a356225eb0547
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2614125
Tested-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>