metrics: Add Platform.UnaggregatedUsageTime metric
The Platform.DailyUseTime metric has many counterintuitive limitations;
most notably the usage time it reports will almost always be on a
different day than the day on which it is reported. Add a new metric,
typically reported every 5 minutes, that can be summed over a period to
give a more accurate view of usage during that period.

BUG=b:176848038
TEST=Push metrics_daemon changes, verify metric appears \
     in chrome://histograms

Change-Id: Ieaf9854c00ccd5f3d2ec215eb412b7eb18de9574
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2615988
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Michael Irani <michaelirani@google.com>