ml: Handle PREFETCH from Soda as a partial, and updated loading test.
Turns out that sometimes soda, in IME mode, still sends Prefetch - so
we're going to handle them like partials for now for convenience. This
is an odd choice to make but we'll honor it here as is.

Update the Load test to ensure that this fails in current environments
- to reflect the reality of the test.

BUG=b:160550533
TEST=end to end test, and unittest execution with USE=ondevice_speech

Change-Id: Iefce66237e0270f8f247c8e938cf9044867eddc0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2626549
Tested-by: Rob Schonberger <robsc@chromium.org>
Commit-Queue: Rob Schonberger <robsc@chromium.org>
Reviewed-by: Andrew Moylan <amoylan@chromium.org>