vm_tools: Allow vmc to boot dlc-based VMs
Concierge has support for booting VMs based on a dlc-id, which is used
to determine their rootfs and kernel paths.  This is the only way of
booting certain classes of VM.

We enable VMC to exercise this functionality, which means it can now be
used to boot dlc-enabled VMs.

BUG=b:177367054
TEST=vmc start --dlc-id=termina-dlc termina

Change-Id: I1f3deea1abcac4b7716abac8fb5dd0ccd1e3b998
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2626561
Tested-by: Nic Hollingum <hollingum@google.com>
Reviewed-by: Fergus Dall <sidereal@google.com>
Reviewed-by: Mike Gerow <gerow@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Commit-Queue: Nic Hollingum <hollingum@google.com>