installer: Manage the cache when copying with "dd"
During "chromeos-install" we end up copying massive amounts of data
from one block device to another.  While we're doing this copy, Linux
needs to temporarily read the data into RAM.  Linux doesn't actually
know what we're doing with this data and doesn't know if we're done
with data or if we're going to ask for it again, so it tries to cache
everything it reads.  As long as we've got lots of RAM, Linux doesn't
have any problems.  However, if once we start running low on RAM Linux
will start to throw things out of its cache.  If we don't give Linux
any hints, it might make the right decision or it might not.

This was a really big problem on "low"-RAM / large rootfs devices
before <https://crrev.com/c/2330304>.  Specifically, we saw problems
on 4GB RAM trogdor devices.  Those devices have a 4GB rootfs
partition, though only 2.3 GB is used right now.  As seen in that
change description, chromeos-install ends up copying the rootfs from
one place in the source to two places in the destination.  If Linux
caches things well then we only need to read from the source once.  If
Linux does a bad job, we need to read things from the source twice.
Since the source is often a slow USB disk, we want to make sure Linux
can do a good job.  After that CL, Linux did a much better job.

However, even after that CL, we've still seen cases were Linux makes
the wrong choice of what to cache and what to throw away.
Specifically, it's seen that some of the rootfs copies are still
slower than they ought to be.  We can do better.

One way we can do better is to split the copy into more chunks.  If
we're copying smaller chunks of the rootfs at a time then Linux
appears much more likely to keep the relevant data in the cache.
Switching from 4 to 8 chunks made the 4GB trogdor devices install 13
seconds faster (averaging over 10 tests each with a slow USB disk as a
source) and avoided instances of chunks installing slower.  However,
increasing the number of chunks appears to have a small perf hit.  It
slowed an 8GB system down by 6 seconds (averaging over 10 tests).
This is consistent with measurements done in
<https://crrev.com/c/2330304>.

Another way to do better is to give hints to Linux about how it should
manage its cache.  Specifically, we can tell Linux that it should
never need to keep the output data in cache.  We can also tell Linux
when it shouldn't bother caching the input data and when it should.
The "dd" command allows you to do this.  After adding these hints:
* 4GB trogdor systems w/ a slow USB disk improved by 20 seconds
  compared to the original code.
* 8GB trogdor systems installed at the same speed.

BUG=chromium:1160096
TEST=chromeos-install is faster

Disallow-Recycled-Builds: test-failures
Change-Id: If6dfc18458b4934e9c81be6cd904f052ec7ba31d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2622585
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>