typecd: Add MockPort for unit tests
Add a MockPort class which derives from a regular Port class. With this,
we can control the behaviour of the class (which modes it supports,
whether cable discovery is complete, etc.) without having to set up an
entire fake sysfs directory and really adding Cable and Partner objects.
This in turn allows to test out various scenarios of the RunModeEntry()
code.

Make the relevant changes to the Port class to allow the MockPort to be
defined.

Also, add a simple unit test that uses a MockPort which can only enter
DP, and validate that the ECUtil function is called correctly.

BUG=b:168653671
TEST=1. Check that all unit tests still pass.
     2. Build and boot on a voxel. Validate that partner, cable and alt
     mode hotplug removal events are being generated as expected in
     typecd.log

Change-Id: I1196144582ed2ee81b9c6fed0db0c1cf7ead02f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2632226
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Tested-by: Prashant Malani <pmalani@chromium.org>
Commit-Queue: Prashant Malani <pmalani@chromium.org>