typecd: Convert enum to enum class
enum class is preferred to plain enums. So, update the enum definition
to make it an enum class. While we are here, update the names of the
constants to avoid repetition of the enum name in the individual enum
values, and add a convenience function to print out the mode as a
string.

This CL does not introduce any functional change.

BUG=None
TEST=Ensure all unit tests still pass.

Change-Id: I23c28c3268652b764f527b9a9891b0494c992a33
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2632405
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Commit-Queue: Prashant Malani <pmalani@chromium.org>
Tested-by: Prashant Malani <pmalani@chromium.org>